Added config.py, check if already updated
Added config.py file to easily change some parameters.

Added check for when a show was last updated, so fewer requests are made. This also takes care of shows that appear in multiple genres.

Also fixed some string formatting so that only a single underscore is used at a time for file path names.